requirements.txt
pillow
sentence-transformers
transformers
pandas
pyarrow
huggingface_hub[hf_xet]
accelerate
torch
scikit-learn
numpy
requests
demoji
swifter[notebook]
reverse_geocoder
python-dotenv
uvicorn
fastapi
pydantic